requirements.txt
Pillow==12.0.0
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.7.2
streamlit==1.51.0
timm==1.0.22
tqdm==4.67.1
torch==2.9.1+cpu
torchvision==0.24.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
